backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
PyJWT==2.9.0
bcrypt==4.2.0
reportlab==4.2.2
python-dotenv==1.0.1

email-validator==2.2.0

pydantic-settings==2.5.2
email-validator==2.2.0
python-dotenv==1.0.1

PyJWT==2.9.0
